Update to CAPU v0.19.0
CAPU 0.19.0 provides bug fixes and improvements listed here:
https://github.com/bmwcarit/capu/releases


git-svn-id: https://svn.apache.org/repos/asf/etch/trunk@1679585 13f79535-47bb-0310-9956-ffa450edef68

+++ binding-cpp/runtime/include/common/EtchSocket.h
@@ -151,7 +151,7 @@ public:
    *         ETCH_SOCKET_ESOCKET if the socket is not created
    *         ETCH_ERROR otherwise
    */
-  status_t getLingerOption(capu::bool_t& isLinger, capu::int32_t& linger);
+  status_t getLingerOption(capu::bool_t& isLinger, capu::uint16_t& linger);
 
   /**
    * Get no delay option
